EP 22 - The principle of least privilege

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -30,15 +30,11 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
-            // policy
-            $this->isAble('store', Ticket::class);
-
+        if($this->isAble('store', Ticket::class)){
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
+
+        return $this->error('You are not authorized to update that resource', 401);
     }
 
     /**
@@ -66,16 +62,14 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // policy
-            $this->isAble('update', $ticket);
+            if($this->isAble('update', $ticket)){
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
     }
 
@@ -85,13 +79,12 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // policy
-            $this->isAble('replace', $ticket);
+            if($this->isAble('replace', $ticket)){
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
-
-            return new TicketResource($ticket);
-
+            return $this->error('You are not authorized to replace that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found', 404);
         }
@@ -105,12 +98,12 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // policy
-            $this->isAble('delete', $ticket);
+            if($this->isAble('delete', $ticket)){
+                $ticket->delete();
+                return $this->ok('Ticket successfully deleted.');
+            }
 
-            $ticket->delete();
-
-            return $this->ok('Ticket successfully deleted.');
+            return $this->error('You are not authorized to delete that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found', 404);
         }
